cross-mount: move under commands/builtin/generic/crossmount, split by command, impl in modules
Relocate the cross-mount layer beside the generics it delegates to. Split the
old 'transfer' (cp/mv/diff/cmp) into transfer.py (cp/mv) and compare.py
(diff/cmp); aggregate -> read.py; adapter -> ops.py; router -> route.py.
Every __init__.py is re-exports only; all implementation lives in named
modules.
