+++ python/mirage/commands/builtin/generic/crossmount/ops.py
